chore: move recent date range tests to WebOverview and WebStats test files
