+++ posthog/hogql_queries/web_analytics/test/test_web_overview.py
@@ -1,4 +1,4 @@
-from datetime import UTC, datetime
+from datetime import UTC, datetime, timedelta
 from typing import Optional
 from zoneinfo import ZoneInfo
 
@@ -974,3 +974,60 @@ class TestWebOverviewQueryRunner(ClickhouseTestMixin, APIBaseTest):
         )
         runner = WebOverviewQueryRunner(team=self.team, query=query)
         assert runner.query_date_range._interval == expected_interval
+
+    @parameterized.expand(
+        [
+            # (name, hours_delta, explicit_date_to, expected_is_recent)
+            ("1_hour_no_explicit_date_to", 1, False, True),
+            ("6_hours_no_explicit_date_to", 6, False, True),
+            ("7_hours_no_explicit_date_to", 7, False, False),
+            ("24_hours_no_explicit_date_to", 24, False, False),
+            ("1_hour_with_explicit_date_to", 1, True, False),
+            ("6_hours_with_explicit_date_to", 6, True, False),
+        ]
+    )
+    def test_is_recent_relative_date_range(self, _name, hours_delta, explicit_date_to, expected_is_recent):
+        now = datetime(2025, 1, 31, 12, 0, 0, tzinfo=UTC)
+        date_from = now - timedelta(hours=hours_delta)
+
+        query = WebOverviewQuery(
+            dateRange=DateRange(date_to=now.isoformat() if explicit_date_to else None),
+            properties=[],
+        )
+
+        runner = MagicMock()
+        runner.query = query
+        runner.team = self.team
+        runner.modifiers = HogQLQueryModifiers(useWebAnalyticsPreAggregatedTables=True, convertToProjectTimezone=False)
+
+        mock_date_range = MagicMock()
+        mock_date_range.date_from.return_value = date_from
+        mock_date_range.date_to.return_value = now
+        runner.query_date_range = mock_date_range
+        runner.query_compare_to_date_range = None
+
+        builder = WebOverviewPreAggregatedQueryBuilder(runner)
+        self.assertEqual(builder._is_recent_relative_date_range(), expected_is_recent)
+
+    def test_can_use_preaggregated_tables_rejects_recent_relative_date_range(self):
+        now = datetime(2025, 1, 31, 12, 0, 0, tzinfo=UTC)
+        date_from = now - timedelta(hours=3)
+
+        query = WebOverviewQuery(
+            dateRange=DateRange(),  # No explicit date_to means query ends at "now"
+            properties=[],
+        )
+
+        runner = MagicMock()
+        runner.query = query
+        runner.team = self.team
+        runner.modifiers = HogQLQueryModifiers(useWebAnalyticsPreAggregatedTables=True, convertToProjectTimezone=False)
+
+        mock_date_range = MagicMock()
+        mock_date_range.date_from.return_value = date_from
+        mock_date_range.date_to.return_value = now
+        runner.query_date_range = mock_date_range
+        runner.query_compare_to_date_range = None
+
+        builder = WebOverviewPreAggregatedQueryBuilder(runner)
+        self.assertFalse(builder.can_use_preaggregated_tables())
